[WIP] move parse_input and generate_alternate_ids to util
